+++ src/pages/sales/POS_partials/POS_Customer_Edit_Sale.tsx
@@ -1,0 +1,3 @@
+const POS_Customer_Edit_sale = () => <></>
+
+export default POS_Customer_Edit_sale;
