+++ src/pages/sales/POS_partials/VSE_Return.tsx
@@ -1,0 +1,3 @@
+const VSE_Return = () => <></>
+
+export default VSE_Return;
