+++ src/pages/sales/POS_partials/VSE_loadout.tsx
@@ -1,0 +1,3 @@
+const VSE_Loadout = () => <></>
+
+export default VSE_Loadout
